requirements.txt
alembic==1.13.2
APScheduler==3.10.4
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.2.2
Werkzeug==3.0.3
zipp==3.19.2
